remove old buttons and fix default button
