--- src/button/button-loading.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-import loading from './loading.svg';
-
-export default styled(loading)`
-  padding-left: 8px;
-`;
