--- src/button/button-warning.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import colors from '../colors';
-import Button from './style';
-
-export default styled(Button)`
-  color: ${colors.white};
-  background-color: ${colors.warning};
-  border: 1px solid ${colors.warning};
-`;
