+++ src/button/index.js
@@ -1,21 +1,26 @@
-import ButtonPrimary from './button-primary';
-import ButtonSecondary from './button-secondary';
-import ButtonWhite from './button-white';
-import ButtonPill from './button-pill';
-import ButtonWarning from './button-warning';
-import ButtonOcean from './button-ocean';
-import ButtonDark from './button-dark';
-import Button from './button';
-import LoadingIcon from './button-loading';
+import styled from 'styled-components';
+import colors from '../colors';
+import ButtonStyled from './style';
+import PropTypes from 'prop-types';
 
-export {
-  Button,
-  ButtonPrimary,
-  ButtonSecondary,
-  ButtonWhite,
-  ButtonPill,
-  ButtonWarning,
-  ButtonOcean,
-  ButtonDark,
-  LoadingIcon
+const Button = styled(ButtonStyled)`
+  color: ${colors.white};
+  background-color: ${colors.coral};
+  border: 1px solid ${colors.coral};
+`;
+
+Button.propTypes = {
+  color: PropTypes.oneOf(['primary', 'secondary', 'warning', 'success']),
+  loading: PropTypes.bool,
+  size: PropTypes.oneOf(['sm', 'md', 'lg']),
+  variant: PropTypes.oneOf(['flat', 'outlined', 'text'])
 };
+
+Button.defaultProps = {
+  color: 'primary',
+  loading: false,
+  size: 'md',
+  variant: 'flat'
+};
+
+export default Button;
