+++ stories/button.stories.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import Button from '../src/button';
+
+storiesOf('Buttons', module).add('Button', () => <Button>Hello Button</Button>);
